requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
pandas==2.0.0
psycopg2==2.9.5
pytest==7.2.0
pytest-mock==3.10.0
pyyaml==6.0.0
requests==2.28.1
SQLAlchemy==1.4.44